requirements.txt
# Core
streamlit==1.45.1
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.6.1
scipy==1.15.2
plotly==6.1.1
matplotlib==3.10.1
seaborn==0.13.2
nbformat==5.10.4
openpyxl==3.1.5
wordcloud==1.9.4
nltk==3.9.1

# LDA
gensim==4.3.3
pyLDAvis==3.4.1

# Embeddings & dim-reduction
sentence-transformers==4.1.0
umap-learn==0.5.7
hdbscan==0.8.40

# BERTopic
bertopic==0.16.4

# Turftopic
turftopic==0.9.2

# BunkaTopics
bunkatopics==0.48

# API-based embeddings (optional — install only the ones you need)
openai==1.82.0
voyageai==0.3.2
google-genai==1.14.0

# STM via R (optional — skip if R is not installed)
# rpy2>=3.5.0
